[Refactoring] implemented Feedback Pt.1

+++ mlir/include/mlir/Analysis/Presburger/Simplex.h
@@ -375,7 +375,7 @@ private:
   ///
   /// \returns the separation type, IneqType::SEPARATE, IneqType::ADJ_EQ, or
   /// IneqType::ADJ_INEQ.
-  IneqType separationType(size_t row);
+  IneqType separationType(unsigned row);
 
   /// Checks that \p unknown is neither a redundant row or a dead column
   ///
